MASTER: Better Google Maps Intergration/Nutrition/FruitBowls
-Version 1.4-
In this new update a lot has changed!
Google Maps: We added a pin so that once a user comes onto the website, then the user already has a marker. Just push get directions and they're off! This is better than older version because there was no marker.

Nutrition: We added a new nutrition section! In this section, we are going to add everything that the consumer would need; carbs, calories, ingrediants etc.
Fruitbowls: Started to add some sample examples as to how it should look like!