MASTER: Critical Changes/Woking Smoothie Bowls Menu!
-Version 1.5-
Critical Changes: In version 1.5 we found a critical bug in the website. The menu section of the navbar was supposed to say 'Smoothie Bowls' instead of 'Fruit Bowls'. Smoothie Bowls are the intended target, Fruit Bowls is where there's cut up fruit in a bowl.

Working Smoothie Bowls Menu: We have started development of the 'Smoothie Bowl' menu. The background of each menu item includes a gradient color of what the 'Smoothie Bowl' is, menu entry title, along with a description.