WR: Carousel/Theme Change
-Version 2.3-
We have a working carousel!
We have some sample images! All copyright but just intended to show off skill!
We changed the theme from black/green to white/black!

Problems: Where are we going to fit the about me page?

Also need to change the smoothie img background to carousel!