WR: Coupon Foundation/Extended Menu
-Version 2.5.5-
We have added groundwork for the extended menu!
Some people want more than 5 options so we made it so that the user can have more options avaikable to them if they really want to see the rest of the menu!
Also we added a coupons section where we can entice people to come with some well... COUPONS! But we hid them so only the hardcore get them!