Hide Link Via Note
This does sound a little bit weird on what this is. Because we are still working on the coupons section of the website, it's important to keep the expieremental version of the website, out of the consumers hands, so they can have the best expierence, and we can develop faster!