Upadting Privacy Policy/Afterpost Done
Hello there!
In this commit we have added exciting new changes that I hope you, and the consumers would enjoy!
When you click the submit button on the form, we added that the website gets redirected to the afterpost page.
Also i'm working on the privacy policy so consumers feel good about sending their information!