Setup Juices/Rename Protein Blends, Shakes.
So there are two sections.
1) Setup juices.
I have went into the menu page and created all of the listings of the juices that Berry Smooth Offers! There are no featured products since not many people actually buy the juices that Berry Smooth provides!

2) Rename Protein Blends, Shakes.
This one is more of my fault on why we're going to make the entire section of this menu Protein Shakes instead of Blends. So, on each menu entry, I put Shakes, NOT blends. Zach's Drink was the only exception, so I just decided to change the Zach's Drink and rename the menu section Shakes!