Setup Build Your Own.
NO CONTENT IS ADDED AT THIS STAGE!
The setup for the build your own section is now complete. We added the layout for the bowls/smoothies and pricing.
Now all I have to do is add the content. What I do have to figure out is where to add extras.