Nav Change - Main Pages For Now.
In this new commit we changed the nav location of the menu. The current commit has the menu BELOW the about us section. We are going to delete products for non index pages, and have the menu as the SECOND link, behind home link of course.
FOR MAIN PAGES:
THIS INCLUDES AND IS LIMITED TO THESE PAGES ONLY AS OF THIS COMMIT!
- Contact Page
- Menu (Overview NOT PRODUCT PAGES)
- Index/Home Page