Small Changes For BETA
The website is now in a beta stage.
Berrysmooth.000webhostapp.com/beta 
Only accessible via link. This shows what we have done and what we need to do. We need to still create the about me page, add pictures, and on top of all of tat, redesign the contact page to make it a lot nicer.