Redesign Contact Page - Icons WIP.
The contact page is set for a beta release, but the page just isn't good enough for public. The icons could be better, but i'm currently looking at the best location to add them would be... They just don't fit anywhere. I might just add them on the footer and redesign footer.