Fixed iPad Scaling Issue.
The iPad version of the website was not very good to begin with. It was just a smaller version of the desktop version, but with scaling issues. I did go back and changed the scaling for the MD or Medium sized devices and it should be a lot better. There is an issue on the iPad Pro where the menu item "Yogurt and Parfeit" (spelled incorrectly), is displayed incorrectly, because the item name is on 2 lines.